INR currency added

+++ app/src/main/java/com/example/udhaarmanager/model/FireStoreModel.kt
@@ -1,13 +1,12 @@
 package com.example.udhaarmanager.model
 
 import android.os.Parcelable
-import kotlinx.android.parcel.Parceler
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class FireStoreModel(
     var title: String? = null,
-    var amount: Long? = null,
+    var amount: Double? = null,
     var transactionType: String? = null,
     var tag: String? = null,
     var borrowDate: String? = null,
